+++ CHANGELOG.md
@@ -2,9 +2,13 @@
 
 ## [Unreleased]
 
-### Added
+### Changed
 
 - [TD-1824] Bump td-perms version to fix relations key
+
+### Fixed
+
+- [TD-1749] Added business concept current version id to search response for relations with business_concept as source/target
 
 ## [2.20.0] 2019-05-27